feat: setup eslint and prretier
