--- src/App.js
@@ -1,9 +1,0 @@
-function App() {
-  return (
-    <div className="App">
-      hello
-    </div>
-  );
-}
-
-export default App;
