+++ src/App.jsx
@@ -1,0 +1,5 @@
+function App() {
+  return <div className="App">hello</div>;
+}
+
+export default App;
